Censored time series simulation application
App makes a basic simulation for the estimation of the right-censored times series with the penalized spline. In app, sample size, spatial variation, censoring level, the coefficient of the AR(1) process and smoothing parameter can be selected by a user. Also, estimated curve is presented in the Figure and bottom of the paper MSE value of the estimation is illustrated.